feat: dock section drag, modal panic chord, and legibility prism glass
The Dock's transient section now aggregates one tile per running
application like the pinned strip, and tiles drag across the section
divider: dropping a transient tile inside the pinned strip pins it at
the previewed slot, dropping a pinned tile past the divider unpins it,
and the strip reflows around the dragged tile so the landing position
stays visible. Hover preview cards no longer come up blank for windows
fully covered on the desktop: the preview path reads the
preview-dedicated surface lists instead of the culled physical-desktop
lists.

High-priority modal dialogs (the capability, confirmation, secret, and
app-picker consent prompts and the low-battery alert) no longer dismiss
on outside clicks; they leave only through their buttons,
Escape/Enter, or the new compositor-owned Ctrl+Alt+Escape panic chord
that forces every open modal to its safe-negative exit. The built-in
four-finger vertical swipe opens the command panel again, reversing
the ADR-0116 rebind to the overview (ADR-0119).

The compositor's liquid-glass pass moved from flux::LiquidGlass* to
the new prism material library (Optics ADR-0063), gaining per-body
material overrides and the backdrop-statistics reduction. Context
menus and tooltips consume it through new legibility-tuned glass roles
(ADR-0120): stronger frost, adaptive tint, damped backdrop chroma, and
plate polarity pinned against the text tone, with the Dock and
launcher menus relaxing the tint over measured-calm backdrops.
Backdrop caching split into capture-side and material-side keys, so
opacity fades and adaptation steps rebuild only the composite;
glass_tint joins the material key, so a color-scheme flip no longer
leaves a stale composite.

Design tokens consolidated in aegis-design: a chrome type scale, the
scene palette, the dock palette, emphasis colors, and chip/cell/
application radii, read by chrome components, the modal prompts, and
aegis-lock.

+++ crates/aegis-lock/src/preview.rs
@@ -282,14 +282,14 @@ impl PreviewApp {
         );
         self.logical_size = (size.0.max(1), size.1.max(1));
         let result = if let Some(render) = &mut self.render {
-            render.resize(self.logical_size, self.scale)
+            render.resize(self.logical_size, self.scale as f32)
         } else {
             self.graphics
                 .create_surface(
                     connection,
                     self.window.wl_surface(),
                     self.logical_size,
-                    self.scale,
+                    self.scale as f32,
                 )
                 .map(|render| self.render = Some(render))
         };
@@ -520,7 +520,7 @@ impl CompositorHandler for PreviewApp {
         self.scale = factor.max(1);
         self.window.wl_surface().set_buffer_scale(self.scale);
         if let Some(render) = &mut self.render
-            && let Err(error) = render.resize(self.logical_size, self.scale)
+            && let Err(error) = render.resize(self.logical_size, self.scale as f32)
         {
             self.fail(error.to_string());
             return;
